refactor: move bitcoin-core-client utility methods to tests folder; exclude test files from build output
