+++ test/integration/parachain/staging/sequential/issue.partial.ts
@@ -5,7 +5,6 @@ import {
     CollateralCurrencyExt,
     currencyIdToMonetaryCurrency,
     DefaultInterBtcApi,
-    getIssueRequestsFromExtrinsicResult,
     InterBtcApi,
     InterbtcPrimitivesVaultId,
     IssueStatus,
@@ -26,8 +25,9 @@ import {
     PARACHAIN_ENDPOINT,
     ESPLORA_BASE_PATH,
 } from "../../../../config";
-import { BitcoinCoreClient } from "../../../../../src/utils/bitcoin-core-client";
-import { issueSingle } from "../../../../../src/utils/issueRedeem";
+import { getIssueRequestsFromExtrinsicResult } from "../../../../utils/issue-redeem";
+import { BitcoinCoreClient } from "../../../../utils/bitcoin-core-client";
+import { issueSingle } from "../../../../utils/issue-redeem";
 import { newVaultId, WrappedCurrency } from "../../../../../src";
 import {
     getCorrespondingCollateralCurrenciesForTests,
